Attack option: number of workers
In the context of a perpetual attack, where the duration of the attack
is set to a very long time, scheduling massive amounts of go-routines
isn't very efficient. This change set enables the specification of the
concurrency level of the attack by number of workers.
The default `0` matches the old behaviour of using one worker per hit
(request).