Attacker: Close response body in error conditions
With 05d58ec, a regression was introduced where the response body wasn't
closed if an error was returned while trying to read it.
This commit fixes this regression. Cheers to @Mistobaan for spotting it.

I wasn't using defer r.Body.Close() before with the assumption that
it would be better to close the body as early as possible in order to
increase TCP socket reuse. Thinking again, I now see that as a premature
optimisation which costed more in regressions than it benefited the
actual behaviour of this piece of code.