attack: Read response bodies in full
PR #326 introduced a way to limit the number of bytes read from HTTP responses but.
However, as pointed out by @smartyjohn in #302 (comment), it didn't solve the need
of reading all the response bodies, in order to observe the full network latency timing,
but discard them at the end.

This was an oversight and unintentional, so this change set changes the semantics
of both the `-max-body` flag in the `attack` command and the correspondent
`vegeta.MaxBody` `Attacker` option.

Instead of limiting the number of bytes **read** from the response bodies,
it limits the number of bytes **captured** from response bodies while
always reading the full body from the underlying connection.

Fixes #339