metrics: Add Throughput to Metrics (#420)
* echosrv: Add sleep flag

* metrics: Add Throughput to Metrics

This commit adds a `Throughput` field to the `Metrics` struct which
represents the rate of successful requests per second. It is calculated
by dividing the number of successful requests by the time span in
seconds starting from the beginning of the earliest request in the
result set and ending in the time the very last response arrived.