attack: Benchmarking with -rate=0 and -max-workers=N
This commit allows setting a zero rate which, together with a reasonable
number of `-max-workers`, behaves as independent concurrent users, each
sending a request and waiting for the response before sending the next
one.

This is part of enabling vegeta to work with the Universal Scalability
Law model, where a concurrency unit is defined as an independent
thread/worker/driver as described above (i.e. with backpressure).

Fixes #378