Add the max-connections option
The connections option is misleading, as it let you think that no more
than the given number of connections will be established during the
attack, whereas it does only control the maximum amount of idle connection
in the pool, but keep the maximum number of connection unlimited.

The max-connection sets the maximum number of connections per target
host, which is what most people certainly expect the connections option
does.